requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.3
django-cors-headers==2.1.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.2.0
django-reversion==3.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
pycryptodome==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.10.0
sqlparse==0.3.0
tablib==0.14.0
uritemplate==3.0.0
urllib3==1.25.7
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
